wip: update the "native" version of "EigenTrustSet"
