+++ circuit/src/dynamic_sets/native.rs
@@ -1,6 +1,5 @@
 use crate::{
-	circuit::PoseidonNativeHasher,
-	eddsa::native::{PublicKey, Signature},
+	circuit::{PoseidonNativeHasher, PoseidonNativeSponge},
 	rns::{compose_big_decimal_f, decompose_big_decimal},
 	utils::fe_to_big,
 };
@@ -12,13 +11,31 @@ use itertools::Itertools;
 use num_bigint::{BigInt, ToBigInt};
 use num_rational::BigRational;
 use num_traits::{FromPrimitive, Zero};
-use secp256k1::{ecdsa, Message};
+use secp256k1::{constants::ONE, ecdsa, Message, PublicKey, Secp256k1, SecretKey};
+use sha3::{Digest, Keccak256};
 use std::collections::HashMap;
 
 /// ECDSA public key
 pub type ECDSAPublicKey = secp256k1::PublicKey;
 /// ECDSA signature
 pub type ECDSASignature = ecdsa::RecoverableSignature;
+
+fn keccak256(data: &[u8]) -> Vec<u8> {
+	let mut hasher = Keccak256::new();
+	hasher.update(data);
+	hasher.finalize().to_vec()
+}
+
+fn recover_ethereum_address_from_pk(pk: ECDSAPublicKey) -> Fr {
+	let pk_bytes = pk.serialize_uncompressed();
+	let hashed_pk = keccak256(&pk_bytes[1..]);
+	let address_bytes = &hashed_pk[hashed_pk.len() - 20..];
+
+	let mut address_bytes_array = [0u8; 32];
+	address_bytes_array[..address_bytes.len()].copy_from_slice(address_bytes);
+
+	Fr::from_bytes(&address_bytes_array).unwrap()
+}
 
 /// Attestation submission struct
 #[derive(Clone)]
@@ -48,6 +65,20 @@ impl SignedAttestation {
 	}
 }
 
+impl Default for SignedAttestation {
+	fn default() -> Self {
+		let attestation = AttestationFr::default();
+
+		let s = Secp256k1::signing_only();
+		let msg = attestation.hash().to_bytes();
+		let sk = SecretKey::from_slice(&ONE).unwrap();
+		let signature =
+			s.sign_ecdsa_recoverable(&Message::from_slice(msg.as_slice()).unwrap(), &sk);
+
+		Self { attestation, signature }
+	}
+}
+
 /// Attestation struct
 #[derive(Clone, Debug)]
 pub struct AttestationFr {
@@ -74,38 +105,23 @@ impl AttestationFr {
 	}
 }
 
+impl Default for AttestationFr {
+	fn default() -> Self {
+		AttestationFr {
+			about: Fr::default(),
+			key: Fr::default(),
+			value: Fr::default(),
+			message: Fr::default(),
+		}
+	}
+}
+
 /// Witness structure for proving threshold checks
 pub struct ThresholdWitness<const NUM_LIMBS: usize> {
 	threshold: Fr,
 	is_bigger: bool,
 	num_decomposed: [Fr; NUM_LIMBS],
 	den_decomposed: [Fr; NUM_LIMBS],
-}
-/// Opinion info of peer
-#[derive(Debug, Clone)]
-pub struct Opinion<const NUM_NEIGHBOURS: usize> {
-	/// Signature of opinion
-	pub sig: Signature,
-	/// Hash of opinion message
-	pub message_hash: Fr,
-	/// Array of real opinions
-	pub scores: Vec<(PublicKey, Fr)>,
-}
-
-impl<const NUM_NEIGHBOURS: usize> Opinion<NUM_NEIGHBOURS> {
-	/// Constructs the instance of `Opinion`
-	pub fn new(sig: Signature, message_hash: Fr, scores: Vec<(PublicKey, Fr)>) -> Self {
-		Self { sig, message_hash, scores }
-	}
-}
-
-impl<const NUM_NEIGHBOURS: usize> Default for Opinion<NUM_NEIGHBOURS> {
-	fn default() -> Self {
-		let sig = Signature::new(Fr::zero(), Fr::zero(), Fr::zero());
-		let message_hash = Fr::zero();
-		let scores = vec![(PublicKey::default(), Fr::zero()); NUM_NEIGHBOURS];
-		Self { sig, message_hash, scores }
-	}
 }
 
 /// Dynamic set for EigenTrust
@@ -114,8 +130,8 @@ pub struct EigenTrustSet<
 	const NUM_ITERATIONS: usize,
 	const INITIAL_SCORE: u128,
 > {
-	set: Vec<(PublicKey, Fr)>,
-	ops: HashMap<PublicKey, Opinion<NUM_NEIGHBOURS>>,
+	set: Vec<(Fr, Fr)>,
+	ops: HashMap<Fr, Vec<Fr>>,
 }
 
 impl<const NUM_NEIGHBOURS: usize, const NUM_ITERATIONS: usize, const INITIAL_SCORE: u128>
@@ -123,130 +139,120 @@ impl<const NUM_NEIGHBOURS: usize, const NUM_ITERATIONS: usize, const INITIAL_SCO
 {
 	/// Constructs new instance
 	pub fn new() -> Self {
-		Self {
-			set: vec![(PublicKey::default(), Fr::zero()); NUM_NEIGHBOURS],
-			ops: HashMap::new(),
-		}
+		Self { set: vec![(Fr::zero(), Fr::zero()); NUM_NEIGHBOURS], ops: HashMap::new() }
 	}
 
 	/// Add new set member and initial score
 	pub fn add_member(&mut self, pk: PublicKey) {
-		let pos = self.set.iter().position(|&(x, _)| x == pk);
+		let pk_fr = recover_ethereum_address_from_pk(pk);
+
+		let pos = self.set.iter().position(|&(x, _)| x == pk_fr);
 		// Make sure not already in the set
 		assert!(pos.is_none());
 
-		let first_available = self.set.iter().position(|&(x, _)| x == PublicKey::default());
+		let first_available = self.set.iter().position(|&(x, _)| x == Fr::zero());
 		let index = first_available.unwrap();
 
 		// Give the initial score.
 		let initial_score = Fr::from_u128(INITIAL_SCORE);
-		self.set[index] = (pk, initial_score);
+		self.set[index] = (pk_fr, initial_score);
 	}
 
 	/// Remove the member and its opinion
 	pub fn remove_member(&mut self, pk: PublicKey) {
+		let pk = recover_ethereum_address_from_pk(pk);
+
 		let pos = self.set.iter().position(|&(x, _)| x == pk);
 		// Make sure already in the set
 		assert!(pos.is_some());
 
 		let index = pos.unwrap();
-		self.set[index] = (PublicKey::default(), Fr::zero());
+		self.set[index] = (Fr::zero(), Fr::zero());
 
 		self.ops.remove(&pk);
 	}
 
 	/// Update the opinion of the member
-	pub fn update_op(&mut self, from: PublicKey, op: Opinion<NUM_NEIGHBOURS>) {
-		let pos_from = self.set.iter().position(|&(x, _)| x == from);
+	pub fn update_op(&mut self, from: ECDSAPublicKey, op: Vec<SignedAttestation>) -> Fr {
+		let from_pk = recover_ethereum_address_from_pk(from);
+		let pos_from = self.set.iter().position(|&(x, _)| x == from_pk);
 		assert!(pos_from.is_some());
 
-		self.ops.insert(from, op);
-	}
+		let mut scores = vec![Fr::zero(); NUM_NEIGHBOURS];
+		let mut hashes = Vec::new();
+		for (i, att) in op.iter().enumerate() {
+			let is_default_pubkey = self.set[i].0 == Fr::zero();
 
-	/// Get a specific opinion from a peer
-	pub fn get_op(&self, from: &PublicKey) -> Opinion<NUM_NEIGHBOURS> {
-		self.ops.get(from).cloned().unwrap_or(Opinion::default())
+			if is_default_pubkey {
+				scores[i] = Fr::default();
+				hashes.push(AttestationFr::default().hash());
+			} else {
+				assert!(att.attestation.about == self.set[i].0);
+
+				let recovered = att.recover_public_key().unwrap();
+				assert!(recovered == from);
+
+				scores[i] = att.attestation.value;
+
+				let hash = att.attestation.hash();
+				hashes.push(hash);
+			}
+		}
+
+		self.ops.insert(from_pk, scores);
+
+		let mut sponge_hasher = PoseidonNativeSponge::new();
+		sponge_hasher.update(&hashes);
+		let op_hash = sponge_hasher.squeeze();
+
+		return op_hash;
 	}
 
 	/// Method for filtering invalid opinions
-	fn filter_peers_ops(&self) -> HashMap<PublicKey, Opinion<NUM_NEIGHBOURS>> {
-		let mut filtered_ops: HashMap<PublicKey, Opinion<NUM_NEIGHBOURS>> = HashMap::new();
+	fn filter_peers_ops(&self) -> HashMap<Fr, Vec<Fr>> {
+		let mut filtered_ops: HashMap<Fr, Vec<Fr>> = HashMap::new();
 
 		// Distribute the scores to valid peers
 		for i in 0..NUM_NEIGHBOURS {
 			let (pk_i, _) = self.set[i].clone();
-			if pk_i == PublicKey::default() {
+			if pk_i == Fr::zero() {
 				continue;
 			}
 
-			let mut ops_i = self.ops.get(&pk_i).unwrap_or(&Opinion::default()).clone();
+			let default_ops = vec![Fr::default(); NUM_NEIGHBOURS];
+			let mut ops_i = self.ops.get(&pk_i).unwrap_or(&default_ops).clone();
 
 			// Update the opinion array - pairs of (key, score)
-			//
-			// Example 1:
-			// 	set => [p1, null, p3]
-			//	Peer1 opinion
-			// 		[(p1, 10), (p6, 10),  (p3, 10)]
-			//   => [(p1, 0), (null, 0), (p3, 10)]
-			//
-			// Example 2:
-			// 	set => [p1, p2, null]
-			//	Peer1 opinion
-			// 		[(p1, 0), (p3, 10), (null, 10)]
-			//   => [(p1, 0), (p2, 0),  (p3, 0)]
 			for j in 0..NUM_NEIGHBOURS {
-				let (set_pk_j, _) = self.set[j];
-				let (op_pk_j, _) = ops_i.scores[j].clone();
+				let (pk_j, _) = self.set[j];
 
-				let is_diff_pk_j = set_pk_j != op_pk_j;
-				let is_pk_j_null = set_pk_j == PublicKey::default();
-				let is_pk_i = set_pk_j == pk_i;
+				// Conditions fro nullifying the score
+				// 1. pk_j == 0(null or default key)
+				// 2. pk_j == pk_i
+				let is_pk_j_null = pk_j == Fr::zero();
+				let is_pk_i = pk_j == pk_i;
 
-				// Conditions for nullifying the score
-				// 1. set_pk_j != op_pk_j
-				// 2. set_pk_j == 0 (null or default key)
-				// 3. set_pk_j == pk_i
-				if is_diff_pk_j || is_pk_j_null || is_pk_i {
-					ops_i.scores[j].1 = Fr::zero();
-				}
-
-				// Condition for correcting the pk
-				// 1. set_pk_j != op_pk_j
-				if is_diff_pk_j {
-					ops_i.scores[j].0 = set_pk_j;
+				if is_pk_j_null || is_pk_i {
+					ops_i[j] = Fr::zero();
 				}
 			}
 
 			// Distribute the scores
-			//
-			// Example 1:
-			// 	set => [p1, p2, p3]
-			//	Peer1 opinion
-			// 		[(p1, 0), (p2, 0), (p3, 10)]
-			//   => [(p1, 0), (p2, 0), (p3, 10)]
-			//
-			// Example 2:
-			// 	set => [p1, p2, p3]
-			//	Peer1 opinion
-			//      [(p1, 0), (p2, 0), (p3, 0)]
-			//   => [(p1, 0), (p2, 1), (p3, 1)]
-			let op_score_sum = ops_i.scores.iter().fold(Fr::zero(), |acc, &(_, score)| acc + score);
+			let op_score_sum = ops_i.iter().fold(Fr::zero(), |acc, &score| acc + score);
 			if op_score_sum == Fr::zero() {
 				for j in 0..NUM_NEIGHBOURS {
-					let (pk_j, _) = ops_i.scores[j].clone();
-
+					let (pk_j, _) = self.set[j];
 					let is_diff_pk = pk_j != pk_i;
-					let is_not_null = pk_j != PublicKey::default();
+					let is_not_null = pk_j != Fr::zero();
 
 					// Conditions for distributing the score
 					// 1. pk_j != pk_i
-					// 2. pk_j != PublicKey::default()
+					// 2. pk_j != Fr::zero()
 					if is_diff_pk && is_not_null {
-						ops_i.scores[j] = (pk_j, Fr::from(1));
+						ops_i[j] = Fr::from(1);
 					}
 				}
 			}
-
 			filtered_ops.insert(pk_i, ops_i);
 		}
 
@@ -256,20 +262,19 @@ impl<const NUM_NEIGHBOURS: usize, const NUM_ITERATIONS: usize, const INITIAL_SCO
 	/// Compute the EigenTrust score
 	pub fn converge(&self) -> Vec<Fr> {
 		// There should be at least 2 valid peers(valid opinions) for calculation
-		let valid_peers = self.set.iter().filter(|(pk, _)| pk != &PublicKey::default()).count();
+		let valid_peers = self.set.iter().filter(|(pk, _)| *pk != Fr::zero()).count();
 		assert!(valid_peers >= 2, "Insufficient peers for calculation!");
 
-		let mut filtered_ops: HashMap<PublicKey, Opinion<NUM_NEIGHBOURS>> = self.filter_peers_ops();
+		let filtered_ops: HashMap<Fr, Vec<Fr>> = self.filter_peers_ops();
 
 		let mut ops = Vec::new();
 		for i in 0..NUM_NEIGHBOURS {
 			let (pk, _) = self.set[i];
-			if pk == PublicKey::default() {
+			if pk == Fr::zero() {
 				ops.push(vec![Fr::zero(); NUM_NEIGHBOURS]);
 			} else {
-				let ops_i = filtered_ops.get_mut(&pk).unwrap();
-				let scores = ops_i.scores.iter().map(|&(_, score)| score).collect_vec();
-				ops.push(scores);
+				let scores = filtered_ops.get(&pk).unwrap();
+				ops.push(scores.clone());
 			}
 		}
 
@@ -285,7 +290,7 @@ impl<const NUM_NEIGHBOURS: usize, const NUM_ITERATIONS: usize, const INITIAL_SCO
 			}
 		}
 
-		// By this point we should use filtered_opinions
+		// By this point we should use filtered opinions
 		let mut s: Vec<Fr> = self.set.iter().map(|(_, score)| score.clone()).collect();
 		let mut new_s: Vec<Fr> = self.set.iter().map(|(_, score)| score.clone()).collect();
 		for _ in 0..NUM_ITERATIONS {
@@ -310,20 +315,17 @@ impl<const NUM_NEIGHBOURS: usize, const NUM_ITERATIONS: usize, const INITIAL_SCO
 
 	/// Compute the EigenTrust score using BigRational numbers
 	pub fn converge_rational(&self) -> Vec<BigRational> {
-		let mut filtered_ops: HashMap<PublicKey, Opinion<NUM_NEIGHBOURS>> = self.filter_peers_ops();
+		let mut filtered_ops: HashMap<Fr, Vec<Fr>> = self.filter_peers_ops();
 
 		let mut ops = Vec::new();
 		for i in 0..NUM_NEIGHBOURS {
 			let (pk, _) = self.set[i];
-			if pk == PublicKey::default() {
+			if pk == Fr::zero() {
 				ops.push(vec![BigInt::zero(); NUM_NEIGHBOURS]);
 			} else {
 				let ops_i = filtered_ops.get_mut(&pk).unwrap();
-				let scores = ops_i
-					.scores
-					.iter()
-					.map(|&(_, score)| fe_to_big(score).to_bigint().unwrap())
-					.collect_vec();
+				let scores =
+					ops_i.iter().map(|&score| fe_to_big(score).to_bigint().unwrap()).collect();
 				ops.push(scores);
 			}
 		}
@@ -401,19 +403,15 @@ impl<const NUM_NEIGHBOURS: usize, const NUM_ITERATIONS: usize, const INITIAL_SCO
 mod test {
 	use std::time::Instant;
 
-	use super::{EigenTrustSet, Opinion, ThresholdWitness};
-	use crate::{
-		calculate_message_hash,
-		eddsa::native::{sign, PublicKey, SecretKey},
-		rns::compose_big_decimal,
-		utils::fe_to_big,
-	};
+	use super::*;
+	use crate::{rns::compose_big_decimal, utils::fe_to_big};
 
 	use halo2::halo2curves::{bn256::Fr, ff::PrimeField};
 	use itertools::Itertools;
 	use num_bigint::ToBigInt;
 	use num_rational::BigRational;
 	use rand::thread_rng;
+	use secp256k1::generate_keypair;
 
 	const NUM_NEIGHBOURS: usize = 12;
 	const NUM_ITERATIONS: usize = 10;
@@ -424,22 +422,33 @@ mod test {
 		const NUM_ITERATIONS: usize,
 		const INITIAL_SCORE: u128,
 	>(
-		sk: &SecretKey, pk: &PublicKey, pks: &[PublicKey], scores: &[Fr],
-	) -> Opinion<NUM_NEIGHBOURS> {
+		sk: &SecretKey, pks: &[Option<PublicKey>], scores: &[Fr],
+	) -> Vec<SignedAttestation> {
 		assert!(pks.len() == NUM_NEIGHBOURS);
 		assert!(scores.len() == NUM_NEIGHBOURS);
 
-		let (_, message_hashes) =
-			calculate_message_hash::<NUM_NEIGHBOURS, 1>(pks.to_vec(), vec![scores.to_vec()]);
-		let sig = sign(sk, pk, message_hashes[0]);
+		let sign = Secp256k1::signing_only();
 
-		// let scores = pks.zip(*scores);
-		let mut op_scores = vec![];
+		let mut res = Vec::new();
 		for i in 0..NUM_NEIGHBOURS {
-			op_scores.push((pks[i], scores[i]));
+			if pks[i].is_none() {
+				res.push(SignedAttestation::default())
+			} else {
+				let about = recover_ethereum_address_from_pk(pks[i].clone().unwrap());
+				let key = Fr::one();
+				let value = scores[i].clone();
+				let message = Fr::one();
+				let attestation = AttestationFr::new(about, key, value, message);
+
+				let message = attestation.hash().to_bytes();
+				let signature = sign
+					.sign_ecdsa_recoverable(&Message::from_slice(message.as_slice()).unwrap(), sk);
+
+				let signed_attestation = SignedAttestation::new(attestation, signature);
+				res.push(signed_attestation);
+			}
 		}
-		let op = Opinion::new(sig, message_hashes[0], op_scores.to_vec());
-		op
+		res
 	}
 
 	#[test]
@@ -449,13 +458,12 @@ mod test {
 
 		let rng = &mut thread_rng();
 
-		let sk1 = SecretKey::random(rng);
-		let pk1 = sk1.public();
+		let (_sk, pk) = generate_keypair(rng);
 
-		set.add_member(pk1);
+		set.add_member(pk);
 
 		// Re-adding the member should panic
-		set.add_member(pk1);
+		set.add_member(pk);
 	}
 
 	#[test]
@@ -465,10 +473,9 @@ mod test {
 
 		let rng = &mut thread_rng();
 
-		let sk1 = SecretKey::random(rng);
-		let pk1 = sk1.public();
+		let (_sk, pk) = generate_keypair(rng);
 
-		set.add_member(pk1);
+		set.add_member(pk);
 
 		set.converge();
 	}
@@ -479,11 +486,8 @@ mod test {
 
 		let rng = &mut thread_rng();
 
-		let sk1 = SecretKey::random(rng);
-		let sk2 = SecretKey::random(rng);
-
-		let pk1 = sk1.public();
-		let pk2 = sk2.public();
+		let (_sk1, pk1) = generate_keypair(rng);
+		let (_sk2, pk2) = generate_keypair(rng);
 
 		set.add_member(pk1);
 		set.add_member(pk2);
@@ -497,26 +501,22 @@ mod test {
 
 		let rng = &mut thread_rng();
 
-		let sk1 = SecretKey::random(rng);
-		let sk2 = SecretKey::random(rng);
-
-		let pk1 = sk1.public();
-		let pk2 = sk2.public();
+		let (sk1, pk1) = generate_keypair(rng);
+		let (_sk2, pk2) = generate_keypair(rng);
 
 		set.add_member(pk1);
 		set.add_member(pk2);
 
 		// Peer1(pk1) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[1] = Fr::from_u128(INITIAL_SCORE);
 
-		let op1 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk1, &pk1, &pks, &scores,
-		);
+		let op1 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk1, &pks, &scores);
 
 		set.update_op(pk1, op1);
 
@@ -529,40 +529,35 @@ mod test {
 
 		let rng = &mut thread_rng();
 
-		let sk1 = SecretKey::random(rng);
-		let sk2 = SecretKey::random(rng);
-
-		let pk1 = sk1.public();
-		let pk2 = sk2.public();
+		let (sk1, pk1) = generate_keypair(rng);
+		let (sk2, pk2) = generate_keypair(rng);
 
 		set.add_member(pk1);
 		set.add_member(pk2);
 
 		// Peer1(pk1) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[1] = Fr::from_u128(INITIAL_SCORE);
 
-		let op1 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk1, &pk1, &pks, &scores,
-		);
+		let op1 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk1, &pks, &scores);
 
 		set.update_op(pk1, op1);
 
 		// Peer2(pk2) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
-		scores[1] = Fr::from_u128(INITIAL_SCORE);
+		scores[0] = Fr::from_u128(INITIAL_SCORE);
 
-		let op2 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk2, &pk2, &pks, &scores,
-		);
+		let op2 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk2, &pks, &scores);
 
 		set.update_op(pk2, op2);
 
@@ -574,64 +569,56 @@ mod test {
 		let mut set = EigenTrustSet::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>::new();
 
 		let rng = &mut thread_rng();
-
-		let sk1 = SecretKey::random(rng);
-		let sk2 = SecretKey::random(rng);
-		let sk3 = SecretKey::random(rng);
-
-		let pk1 = sk1.public();
-		let pk2 = sk2.public();
-		let pk3 = sk3.public();
+		let (sk1, pk1) = generate_keypair(rng);
+		let (sk2, pk2) = generate_keypair(rng);
+		let (sk3, pk3) = generate_keypair(rng);
 
 		set.add_member(pk1);
 		set.add_member(pk2);
 		set.add_member(pk3);
 
 		// Peer1(pk1) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[1] = Fr::from_u128(300);
 		scores[2] = Fr::from_u128(700);
 
-		let op1 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk1, &pk1, &pks, &scores,
-		);
+		let op1 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk1, &pks, &scores);
 
 		set.update_op(pk1, op1);
 
 		// Peer2(pk2) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[0] = Fr::from_u128(600);
 		scores[2] = Fr::from_u128(400);
 
-		let op2 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk2, &pk2, &pks, &scores,
-		);
+		let op2 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk2, &pks, &scores);
 
 		set.update_op(pk2, op2);
 
 		// Peer3(pk3) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[0] = Fr::from_u128(600);
 		scores[1] = Fr::from_u128(400);
 
-		let op3 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk3, &pk3, &pks, &scores,
-		);
+		let op3 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk3, &pks, &scores);
 
 		set.update_op(pk3, op3);
 
@@ -644,47 +631,41 @@ mod test {
 
 		let rng = &mut thread_rng();
 
-		let sk1 = SecretKey::random(rng);
-		let sk2 = SecretKey::random(rng);
-		let sk3 = SecretKey::random(rng);
-
-		let pk1 = sk1.public();
-		let pk2 = sk2.public();
-		let pk3 = sk3.public();
+		let (sk1, pk1) = generate_keypair(rng);
+		let (sk2, pk2) = generate_keypair(rng);
+		let (sk3, pk3) = generate_keypair(rng);
 
 		set.add_member(pk1);
 		set.add_member(pk2);
 		set.add_member(pk3);
 
 		// Peer1(pk1) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[1] = Fr::from_u128(300);
 		scores[2] = Fr::from_u128(700);
 
-		let op1 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk1, &pk1, &pks, &scores,
-		);
+		let op1 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk1, &pks, &scores);
 
 		set.update_op(pk1, op1);
 
 		// Peer2(pk2) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[0] = Fr::from_u128(600);
 		scores[2] = Fr::from_u128(400);
 
-		let op2 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk2, &pk2, &pks, &scores,
-		);
+		let op2 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk2, &pks, &scores);
 
 		set.update_op(pk2, op2);
 
@@ -696,65 +677,56 @@ mod test {
 		let mut set = EigenTrustSet::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>::new();
 
 		let rng = &mut thread_rng();
-
-		let sk1 = SecretKey::random(rng);
-		let sk2 = SecretKey::random(rng);
-		let sk3 = SecretKey::random(rng);
-
-		let pk1 = sk1.public();
-		let pk2 = sk2.public();
-		let pk3 = sk3.public();
+		let (sk1, pk1) = generate_keypair(rng);
+		let (sk2, pk2) = generate_keypair(rng);
+		let (sk3, pk3) = generate_keypair(rng);
 
 		set.add_member(pk1);
 		set.add_member(pk2);
 		set.add_member(pk3);
 
 		// Peer1(pk1) signs the opinion
-		// Peer2(pk2) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[1] = Fr::from_u128(300);
 		scores[2] = Fr::from_u128(700);
 
-		let op1 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk1, &pk1, &pks, &scores,
-		);
+		let op1 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk1, &pks, &scores);
 
 		set.update_op(pk1, op1);
 
 		// Peer2(pk2) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[0] = Fr::from_u128(600);
 		scores[2] = Fr::from_u128(400);
 
-		let op2 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk2, &pk2, &pks, &scores,
-		);
+		let op2 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk2, &pks, &scores);
 
 		set.update_op(pk2, op2);
 
 		// Peer3(pk3) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[0] = Fr::from_u128(600);
 		scores[1] = Fr::from_u128(400);
 
-		let op3 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk3, &pk3, &pks, &scores,
-		);
+		let op3 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk3, &pks, &scores);
 
 		set.update_op(pk3, op3);
 
@@ -772,47 +744,41 @@ mod test {
 
 		let rng = &mut thread_rng();
 
-		let sk1 = SecretKey::random(rng);
-		let sk2 = SecretKey::random(rng);
-		let sk3 = SecretKey::random(rng);
-
-		let pk1 = sk1.public();
-		let pk2 = sk2.public();
-		let pk3 = sk3.public();
+		let (sk1, pk1) = generate_keypair(rng);
+		let (sk2, pk2) = generate_keypair(rng);
+		let (sk3, pk3) = generate_keypair(rng);
 
 		set.add_member(pk1);
 		set.add_member(pk2);
 		set.add_member(pk3);
 
 		// Peer1(pk1) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[1] = Fr::from_u128(300);
 		scores[2] = Fr::from_u128(700);
 
-		let op1 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk1, &pk1, &pks, &scores,
-		);
+		let op1 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk1, &pks, &scores);
 
 		set.update_op(pk1, op1);
 
 		// Peer2(pk2) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[0] = Fr::from_u128(600);
 		scores[2] = Fr::from_u128(400);
 
-		let op2 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk2, &pk2, &pks, &scores,
-		);
+		let op2 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk2, &pks, &scores);
 
 		set.update_op(pk2, op2);
 
@@ -837,53 +803,46 @@ mod test {
 
 		let rng = &mut thread_rng();
 
-		let sk1 = SecretKey::random(rng);
-		let sk2 = SecretKey::random(rng);
-		let sk3 = SecretKey::random(rng);
-
-		let pk1 = sk1.public();
-		let pk2 = sk2.public();
-		let pk3 = sk3.public();
+		let (sk1, pk1) = generate_keypair(rng);
+		let (sk2, pk2) = generate_keypair(rng);
+		let (sk3, pk3) = generate_keypair(rng);
 
 		// Peer1(pk1) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[0] = Fr::from_u128(10);
 		scores[1] = Fr::from_u128(10);
 
-		let op1 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk1, &pk1, &pks, &scores,
-		);
+		let op1 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk1, &pks, &scores);
 
 		// Peer2(pk2) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[2] = Fr::from_u128(30);
 
-		let op2 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk2, &pk2, &pks, &scores,
-		);
+		let op2 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk2, &pks, &scores);
 
 		// Peer3(pk3) signs the opinion
-		let mut pks = [PublicKey::default(); NUM_NEIGHBOURS];
-		pks[0] = pk1;
-		pks[1] = pk2;
-		pks[2] = pk3;
+		let mut pks = [None; NUM_NEIGHBOURS];
+		pks[0] = Some(pk1);
+		pks[1] = Some(pk2);
+		pks[2] = Some(pk3);
 
 		let mut scores = [Fr::zero(); NUM_NEIGHBOURS];
 		scores[0] = Fr::from_u128(10);
 
-		let op3 = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-			&sk3, &pk3, &pks, &scores,
-		);
+		let op3 =
+			sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(&sk3, &pks, &scores);
 
 		// Setup EigenTrustSet
 		let mut eigen_trust_set =
@@ -899,10 +858,10 @@ mod test {
 
 		let filtered_ops = eigen_trust_set.filter_peers_ops();
 
-		let final_peers_cnt =
-			eigen_trust_set.set.iter().filter(|&&(pk, _)| pk != PublicKey::default()).count();
-		let final_ops_cnt = filtered_ops.keys().count();
-		assert!(final_peers_cnt == final_ops_cnt);
+		let final_peers_count =
+			eigen_trust_set.set.iter().filter(|&&(pk, _)| pk != Fr::zero()).count();
+		let final_ops_count = filtered_ops.keys().count();
+		assert!(final_peers_count == final_ops_count);
 	}
 
 	fn eigen_trust_set_testing_helper<
@@ -923,22 +882,24 @@ mod test {
 
 		let rng = &mut thread_rng();
 
-		let sks: Vec<SecretKey> =
-			(0..NUM_NEIGHBOURS).into_iter().map(|__| SecretKey::random(rng)).collect();
-		let pks: Vec<PublicKey> = sks.clone().iter().map(|s| s.public()).collect();
+		let keys: Vec<(SecretKey, PublicKey)> =
+			(0..NUM_NEIGHBOURS).into_iter().map(|_| generate_keypair(rng)).collect();
+		let sks: Vec<SecretKey> = keys.iter().map(|(sk, _)| sk.clone()).collect();
+		let pks: Vec<Option<PublicKey>> = keys.iter().map(|(_, pk)| Some(pk.clone())).collect();
 
 		// Add the publicKey to the set
-		pks.iter().for_each(|pk| set.add_member(*pk));
+		pks.iter().for_each(|pk| set.add_member(pk.unwrap()));
 
 		// Update the opinions
 		for i in 0..NUM_NEIGHBOURS {
 			let scores = ops[i].to_vec();
 
 			let op_i = sign_opinion::<NUM_NEIGHBOURS, NUM_ITERATIONS, INITIAL_SCORE>(
-				&sks[i], &pks[i], &pks, &scores,
+				&sks[i], &pks, &scores,
 			);
 
-			set.update_op(pks[i], op_i);
+			let pk_i = pks[i].unwrap();
+			set.update_op(pk_i, op_i);
 		}
 
 		let s = set.converge();
